Firefly Jar functionality
Made the firefly jar functional.
Now changes Lit state when right-clicked.
Can emit particles at a higher light level compared to the firefly bush (10 rather than 7).
When Lit it has a light level of 15.
When not Lit it emits fireflies in a smaller radius than the bush at twice the rate.
(Currently emits no light when not lit, but I want to change this so that it emits a light level of 10 when not lit)

Added Comments in the bush to suggest a better way around collecting fireflies from the bush as currently it works weirdly. Unsure how to implement these changes.