Clifftree Compat and Tag Changes
Added an internal datapack for Cliff Under A Tree compatibility.
Currently, doesn't activate automatically like the Terralith pack.

Biome tags have been changed again.
Tags are now per structure since we have no use for a general swamp tag.
This will allow for us to pick the biomes villages and ruins generate in separately.
For example the Clifftree Bog is too barren for a village but the ruin works fine in it.
Biomes will be added as more generation mods are made compatible.